Phase 4 Wave A: move the admin console into surfaces/admin/
First git-mv batch of the core/+surfaces/ scaffolding named in the plan.
AdminConsole and its 12 panels (assessments, jobs, costs, creation limits,
featured pool, access tokens, public play, suggestions, waitlist) plus
adminApi/adminJobsApi move as one unit — they're already lazy-loaded
together from App.tsx since #1064, and an exhaustive grep of every file
in this cluster confirmed the only outside importer anywhere is
NavHeader.tsx's fetchAdminSummary() call (the header's admin badge/count),
now pointed at the new path. No apps/web module-boundary rule exists yet
(unlike apps/api's buckets), so that one cross-surface import is left as
is rather than forced into a premature core/ extraction.

Caught two things tsc alone couldn't: App.tsx's dynamic
`import('./AdminConsole.js')` lazy-route call (a template string, not a
`from` import, missed by the first grep pass) and three test files'
`vi.mock('./X.js', ...)` calls whose path arguments have to keep
resolving to the same module the component under test actually imports,
or the mock silently stops intercepting anything.

Verified: tsc clean, full apps/web suite (200 files / 1762 tests) green,
repo lint exits 0.
