+++ eslint-rules/comment-prose-baseline.json
@@ -584,6 +584,21 @@
     "apps/web/src/SubmissionStatusView.tsx": 4172,
     "apps/web/src/SuggestionsPanel.test.tsx": 28,
     "apps/web/src/SuggestionsPanel.tsx": 113,
+    "apps/web/src/surfaces/admin/AccessTokensPanel.test.tsx": 75,
+    "apps/web/src/surfaces/admin/AccessTokensPanel.tsx": 185,
+    "apps/web/src/surfaces/admin/adminApi.ts": 223,
+    "apps/web/src/surfaces/admin/AdminConsole.test.tsx": 156,
+    "apps/web/src/surfaces/admin/AdminConsole.tsx": 617,
+    "apps/web/src/surfaces/admin/adminJobsApi.ts": 230,
+    "apps/web/src/surfaces/admin/AdminJobsPanel.test.tsx": 34,
+    "apps/web/src/surfaces/admin/CostsPanel.test.tsx": 120,
+    "apps/web/src/surfaces/admin/CostsPanel.tsx": 217,
+    "apps/web/src/surfaces/admin/CreationLimitsPanel.test.tsx": 45,
+    "apps/web/src/surfaces/admin/CreationLimitsPanel.tsx": 157,
+    "apps/web/src/surfaces/admin/SuggestionsPanel.test.tsx": 28,
+    "apps/web/src/surfaces/admin/SuggestionsPanel.tsx": 113,
+    "apps/web/src/surfaces/admin/WaitlistPanel.test.tsx": 11,
+    "apps/web/src/surfaces/admin/WaitlistPanel.tsx": 164,
     "apps/web/src/telemetry.test.ts": 201,
     "apps/web/src/telemetry.ts": 968,
     "apps/web/src/TelemetryCharts.test.tsx": 26,
