+++ eslint-rules/module-size-baseline.json
@@ -834,6 +834,29 @@
     "apps/web/src/SubmissionStatusView.tsx": 2649,
     "apps/web/src/SuggestionsPanel.test.tsx": 115,
     "apps/web/src/SuggestionsPanel.tsx": 146,
+    "apps/web/src/surfaces/admin/AccessTokensPanel.test.tsx": 181,
+    "apps/web/src/surfaces/admin/AccessTokensPanel.tsx": 185,
+    "apps/web/src/surfaces/admin/adminApi.ts": 491,
+    "apps/web/src/surfaces/admin/AdminAssessmentResolve.tsx": 127,
+    "apps/web/src/surfaces/admin/AdminAssessmentsPanel.test.tsx": 125,
+    "apps/web/src/surfaces/admin/AdminAssessmentsPanel.tsx": 377,
+    "apps/web/src/surfaces/admin/AdminConsole.test.tsx": 256,
+    "apps/web/src/surfaces/admin/AdminConsole.tsx": 274,
+    "apps/web/src/surfaces/admin/AdminJobPreviewModal.tsx": 136,
+    "apps/web/src/surfaces/admin/adminJobsApi.ts": 206,
+    "apps/web/src/surfaces/admin/AdminJobsPanel.test.tsx": 581,
+    "apps/web/src/surfaces/admin/AdminJobsPanel.tsx": 847,
+    "apps/web/src/surfaces/admin/CostsPanel.test.tsx": 193,
+    "apps/web/src/surfaces/admin/CostsPanel.tsx": 171,
+    "apps/web/src/surfaces/admin/CreationLimitsPanel.test.tsx": 398,
+    "apps/web/src/surfaces/admin/CreationLimitsPanel.tsx": 411,
+    "apps/web/src/surfaces/admin/FeaturedPoolPanel.tsx": 97,
+    "apps/web/src/surfaces/admin/PublicPlayPanel.test.tsx": 71,
+    "apps/web/src/surfaces/admin/PublicPlayPanel.tsx": 98,
+    "apps/web/src/surfaces/admin/SuggestionsPanel.test.tsx": 115,
+    "apps/web/src/surfaces/admin/SuggestionsPanel.tsx": 146,
+    "apps/web/src/surfaces/admin/WaitlistPanel.test.tsx": 197,
+    "apps/web/src/surfaces/admin/WaitlistPanel.tsx": 346,
     "apps/web/src/telemetry.test.ts": 285,
     "apps/web/src/telemetry.ts": 273,
     "apps/web/src/TelemetryCharts.test.tsx": 197,
